FELIX-4888 : ServletHandler's are not sorted by longest matching path. Finish error page registry. Start DTO handling (WiP)
git-svn-id: https://svn.apache.org/repos/asf/felix/trunk/http/base@1679973 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/felix/http/base/internal/runtime/dto/ErrorPageDTOBuilder.java
@@ -44,9 +44,9 @@ final class ErrorPageDTOBuilder<T extends ErrorPageDTO> extends BaseServletDTOBu
         return errorPageDTO;
     }
 
-    private long[] getErrorCodes(Collection<Integer> errorCodes)
+    private long[] getErrorCodes(Collection<Long> errorCodes)
     {
-        Iterator<Integer> itr = errorCodes.iterator();
+        Iterator<Long> itr = errorCodes.iterator();
         long[] result = new long[errorCodes.size()];
         for (int i = 0; i < result.length; i++)
         {
